requirements.txt
torch>=2.3
torchvision>=0.18
numpy
pyyaml
opencv-python
albumentations
matplotlib
tqdm
pycocotools
